fixed error in wildcards regarding sorting process
