+++ srcs/execution/wildcards/wildcards_utils.c
@@ -6,12 +6,12 @@
 /*   By: dnikifor <dnikifor@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/22 15:45:50 by dnikifor          #+#    #+#             */
-/*   Updated: 2024/03/26 20:02:49 by dnikifor         ###   ########.fr       */
+/*   Updated: 2024/03/26 20:42:57 by dnikifor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../headers/minishell.h"
-int	sort_array_with_command(char ***arr);
+
 /**
 * @brief	Function that allocates @c `len(arr)` 2d arrays
 * @param	arr pointer to the initial array
@@ -73,6 +73,8 @@ int	fill_temp_array(char ***arr, t_w_cards *wc)
 			ft_free_3d_array(wc->temp_arr, 0);
 			return (SYSTEM_ERROR);
 		}
+		if (array_sorting_process(wc, &i) == MALLOC_ERR)
+			return (MALLOC_ERR);
 	}
 	return (SUCCESS);
 }
@@ -109,50 +111,5 @@ int	allocate_and_fill_expanded_array(t_w_cards *wc)
 		while ((wc->temp_arr)[j][++k])
 			(wc->new_arr)[++i] = (wc->temp_arr)[j][k];
 	}
-	return (SUCCESS);
-}
-
-/**
- * @brief	Function that joins sorted final array with command string
- * @param	arr pointer to the initial array
- * @param	str command string
- * @return	joined array
- */
-char	**str_array_join(char **arr, char *str)
-{
-	char	**joined_arr;
-	int		len;
-	int		i;
-
-	i = -1;
-	len = ft_arrlen((void **)arr);
-	joined_arr = ft_calloc(len + 2, sizeof(char *));
-	if (!joined_arr)
-		return (NULL);
-	joined_arr[0] = str;
-	while (++i < len)
-		joined_arr[i + 1] = arr[i];
-	return (joined_arr);
-}
-
-/**
- * @brief	Function that sorts and then joins final array
- * @param	wc pointer to the structure with local variables initiated in
- * @c `wildcards()`
- * @return	@c `SUCCESS` if no errors occured, @c `MALLOC_ERR` if one of the
- * mallocs failed
- */
-int	sort_and_fill_final_array(t_w_cards *wc)
-{
-	wc->sorted_new_arr = sort_string_arr(wc->new_arr + 1,
-			ft_arrlen((void **)wc->new_arr) - 1);
-	if (!wc->sorted_new_arr)
-	{
-		ft_free_2d_array(wc->new_arr);
-		return (MALLOC_ERR);
-	}
-	wc->final_new_arr = str_array_join(wc->sorted_new_arr, (wc->new_arr)[0]);
-	free(wc->new_arr);
-	free(wc->sorted_new_arr);
 	return (SUCCESS);
 }
