+++ srcs/main.c
@@ -6,7 +6,7 @@
 /*   By: dnikifor <dnikifor@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/21 15:09:34 by dshatilo          #+#    #+#             */
-/*   Updated: 2024/03/22 15:51:16 by dnikifor         ###   ########.fr       */
+/*   Updated: 2024/03/26 20:11:01 by dnikifor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
